requirements.txt
fastapi==0.121.0
uvicorn==0.38.0
requests==2.32.5
beautifulsoup4==4.14.2
python-dotenv==1.2.1
